fix: added clx to dependency
