+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@hookform/resolvers": "^3.9.0",
     "@strategic-dot/components": "^0.0.28",
+    "class-variance-authority": "^0.7.0",
     "@t3-oss/env-nextjs": "^0.10.1",
     "@types/jest-axe": "^3.5.9",
     "framer-motion": "^11.3.8",
