+++ src/components/miscellaneous/blur-image.tsx
@@ -2,7 +2,6 @@
 
 import Image from "next/image";
 import { useState, type ComponentProps } from "react";
-
 import { cn } from "~/lib/utils";
 
 export default function BlurImage(properties: ComponentProps<typeof Image>) {
