+++ src/components/miscellaneous/loading-spinner.tsx
@@ -1,5 +1,4 @@
 import React, { forwardRef } from "react";
-
 import { cn } from "~/lib/utils";
 
 export interface SVGProperties extends React.SVGAttributes<SVGSVGElement> {
